fix: use integrated loading and cleanup components
