--- apps/web/app/environments/[environmentId]/people/Loading.tsx
@@ -1,5 +1,0 @@
-import LoadingTable from "@/components/shared/LoadingTable";
-
-export default function Loading() {
-    return <LoadingTable />
-}
